requirements.txt
pydantic==2.5.0
python-dotenv==1.0.0
neo4j==5.15.0
pytest
pytest-mock
pytest-asyncio
openai>=1.0.0
fastapi
watchdog==3.0.0
pypdf>=3.0.0
uvicorn
jsonlines
google-generativeai
tavily-python
rich
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
gspread
tenacity
tqdm
unidecode
oauth2client
numpy
flatlib
mcp
